fix(render) - refactor rhythm scene
