+++ include/game/components/render/text.h
@@ -9,6 +9,7 @@ namespace Game::Render
     {
         const AssetsRecord *font = nullptr; //font
         std::string text; // text
+        std::string name;
         uint32_t layer = 0;     // Layer to render
         uint32_t order = 0;     // Order of render in layer
     };
